+++ StripePayments/StripePaymentsTests/IntentConfirmationChallengeViewControllerTests.swift
@@ -2,6 +2,8 @@
 //  IntentConfirmationChallengeViewControllerTests.swift
 //  StripePaymentsTests
 //
+//  Created by Joyce Qin on 5/14/26.
+//
 
 import XCTest
 
@@ -42,13 +44,6 @@ class IntentConfirmationChallengeViewControllerTests: XCTestCase {
         }
     }
 
-    func testHandleSuccessCompletionCalledOnce() {
-        var callCount = 0
-        let vc = makeVC { _ in callCount += 1 }
-        vc.handleSuccess()
-        XCTAssertEqual(callCount, 1)
-    }
-
     // MARK: - handleError
 
     func testHandleErrorCallsCompletionWithFailure() {
@@ -63,19 +58,6 @@ class IntentConfirmationChallengeViewControllerTests: XCTestCase {
             return
         }
         XCTAssertEqual(msg, "failed")
-    }
-
-    func testHandleErrorWithUnknownErrorCallsCompletionWithFailure() {
-        var result: Result<Void, Error>?
-        let vc = makeVC { result = $0 }
-        vc.handleError(ChallengeError.unknownError)
-
-        guard case .failure(let error) = result,
-              let challengeError = error as? ChallengeError,
-              case .unknownError = challengeError else {
-            XCTFail("Expected .failure(.unknownError), got \(String(describing: result))")
-            return
-        }
     }
 
     // MARK: - handleReady
@@ -102,4 +84,64 @@ class IntentConfirmationChallengeViewControllerTests: XCTestCase {
         }
     }
 
+    // MARK: - ChallengeError: analyticsErrorType
+
+    func testWebErrorAnalyticsTypeUsesProvidedType() {
+        let error = ChallengeError.webError(message: "msg", type: "arkose_error", code: "1234")
+        XCTAssertEqual(error.analyticsErrorType, "arkose_error")
+    }
+
+    func testNonWebErrorsUseDefaultAnalyticsType() {
+        let underlying = NSError(domain: "test", code: 1)
+        XCTAssertEqual(ChallengeError.userCanceled.analyticsErrorType, "IntentConfirmationChallengeError")
+        XCTAssertEqual(ChallengeError.unknownError.analyticsErrorType, "IntentConfirmationChallengeError")
+        XCTAssertEqual(ChallengeError.navigationFailed(underlying).analyticsErrorType, "IntentConfirmationChallengeError")
+    }
+
+    // MARK: - ChallengeError: analyticsErrorCode
+
+    func testWebErrorAnalyticsCodeUsesProvidedCode() {
+        let error = ChallengeError.webError(message: "msg", type: "type", code: "MY_CODE")
+        XCTAssertEqual(error.analyticsErrorCode, "MY_CODE")
+    }
+
+    func testWebErrorAnalyticsCodeNilDefaultsToUnknown() {
+        let error = ChallengeError.webError(message: "msg", type: "type", code: nil)
+        XCTAssertEqual(error.analyticsErrorCode, "unknown")
+    }
+
+    // MARK: - ChallengeError: additionalNonPIIErrorDetails
+
+    func testWebErrorAdditionalDetailsHasFromBridgeTrue() {
+        let error = ChallengeError.webError(message: "msg", type: "type", code: nil)
+        XCTAssertEqual(error.additionalNonPIIErrorDetails["from_bridge"] as? Bool, true)
+    }
+
+    func testNonWebErrorsAdditionalDetailsHaveFromBridgeFalse() {
+        let underlying = NSError(domain: "test", code: 1)
+        XCTAssertEqual(ChallengeError.userCanceled.additionalNonPIIErrorDetails["from_bridge"] as? Bool, false)
+        XCTAssertEqual(ChallengeError.unknownError.additionalNonPIIErrorDetails["from_bridge"] as? Bool, false)
+        XCTAssertEqual(ChallengeError.navigationFailed(underlying).additionalNonPIIErrorDetails["from_bridge"] as? Bool, false)
+    }
+
+    // MARK: - ChallengeError: errorDescription
+
+    func testWebErrorDescriptionUsesMessage() {
+        let error = ChallengeError.webError(message: "Captcha verification failed", type: "type", code: nil)
+        XCTAssertEqual(error.errorDescription, "Captcha verification failed")
+    }
+
+    func testNavigationFailedDescriptionIncludesUnderlyingError() {
+        let underlying = NSError(domain: "test", code: 1, userInfo: [NSLocalizedDescriptionKey: "Connection lost"])
+        let error = ChallengeError.navigationFailed(underlying)
+        XCTAssertEqual(error.errorDescription, "Navigation failed: Connection lost")
+    }
+
+    func testUserCanceledDescriptionIsNil() {
+        XCTAssertNil(ChallengeError.userCanceled.errorDescription)
+    }
+
+    func testUnknownErrorHasDescription() {
+        XCTAssertEqual(ChallengeError.unknownError.errorDescription, "Unknown error.")
+    }
 }
